AgenticAI_Krish/requirements.txt
langchain
python-dotenv
pymupdf
langchain-community
langchain-text-splitters
langchain-openai
faiss-cpu
langchain-classic
streamlit
langchain-ollama
langchain-groq
fastapi
uvicorn
langserve
sse_starlette
transformers
grandalf
langchain-google-genai